Removing ApiServer from controller.py
