+++ kytos/core/controller.py
@@ -22,6 +22,7 @@ from threading import Thread
 from urllib.request import urlopen
 
 from flask import Flask, request
+from kytos.core.api_server import APIServer
 from kytos.core.buffers import KytosBuffers
 from kytos.core.events import KytosEvent
 from kytos.core.helpers import now, start_logger
@@ -93,48 +94,12 @@ class Controller(object):
         self.log_websocket = LogWebSocket()
         self.console_websocket = ConsoleWebSocket(controller=self)
 
-        self.app = Flask(__name__)
+        self.app = APIServer(__name__)
 
         #: Adding the napps 'enabled' directory into the PATH
         #: Now you can access the enabled napps with:
         #: from napps.<author>.<napp_name> import ?....
         sys.path.append(os.path.join(self.options.napps, os.pardir))
-
-    def register_kytos_routes(self):
-        """Register initial routes from kytos using ApiServer.
-
-        Initial routes are: ['/kytos/status', '/kytos/shutdown']
-        """
-        if '/kytos/status/' not in self.rest_endpoints:
-            self.app.add_url_rule('/kytos/status/', self.status_api.__name__,
-                                  self.status_api, methods=['GET'])
-
-        if '/kytos/shutdown' not in self.rest_endpoints:
-            self.app.add_url_rule('/kytos/shutdown',
-                                  self.shutdown_api.__name__,
-                                  self.shutdown_api, methods=['GET'])
-
-    def register_rest_endpoint(self, url, function, methods):
-        r"""Register a new rest endpoint in Api Server.
-
-        To register new endpoints is needed to have a url, function to handle
-        the requests and type of method allowed.
-
-        Parameters:
-            url (string):        String with partner of route. e.g.: '/status'
-            function (function): Function pointer used to handle the requests.
-            methods (list):      List of request methods allowed.
-                                 e.g: ['GET', 'PUT', 'POST', 'DELETE', 'PATCH']
-        """
-        if url not in self.rest_endpoints:
-            new_endpoint_url = "/kytos{}".format(url)
-            self.app.add_url_rule(new_endpoint_url, function.__name__,
-                                  function, methods=methods)
-
-    @property
-    def rest_endpoints(self):
-        """Return string with routes registered by Api Server."""
-        return [x.rule for x in self.app.url_map.iter_rules()]
 
     def start_log_websocket(self):
         """Start the kytos websocket server."""
@@ -187,6 +152,14 @@ class Controller(object):
         self.started_at = now()
         self.register_kytos_routes()
 
+    def stop_api_server(self):
+        """Method used to send a shutdown request to stop Api Server."""
+        if self.app:
+            urlopen('http://127.0.0.1:8181/kytos/shutdown')
+
+    def register_rest_endpoint(self, *options, **kwargs):
+        self.app.register_rest_endpoint(*options,**kwargs)
+
     def stop(self, graceful=True):
         """Method used to shutdown all services used by kytos.
 
@@ -236,35 +209,6 @@ class Controller(object):
         self.unload_napps()
         self.buffers = KytosBuffers()
         self.server.server_close()
-
-    def stop_api_server(self):
-        """Method used to send a shutdown request to stop Api Server."""
-        if self.app:
-            urlopen('http://127.0.0.1:8181/kytos/shutdown')
-
-    def shutdown_api(self):
-        """Handle shutdown requests received by Api Server.
-
-        This method must be called by kytos using the method
-        stop_api_server, otherwise this request will be ignored.
-        """
-        allowed_host = ['127.0.0.1:8181', 'localhost:8181']
-        if request.host not in allowed_host:
-            return "", 403
-
-        server_shutdown = request.environ.get('werkzeug.server.shutdown')
-        if server_shutdown is None:
-            raise RuntimeError('Not running with the Werkzeug Server')
-        server_shutdown()
-        self.app = None
-        return 'Server shutting down...', 200
-
-    def status_api(self):
-        """Display json with kytos status using the route '/status'."""
-        if self._threads['api_server'].is_alive():
-            return '{"response": "running"}', 201
-        else:
-            return '{"response": "not running"}', 404
 
     def status(self):
         """Return status of Kytos Server.
